feat(vim): Remove swapped mappings for marks
` is more precise and important for `[ and `] .
Also e.g. `a (assuming 'a', 'b', 'c', etc. are the go-to for quick marks, as opposed to letters on the right side of the keyboard).

`a, `b, `c, etc. are uncomfortable to type on English US layout – that's probably why this swap was introduced in the first place.
It is more acceptable on Kinesis Advantage and English internation layout though.
Remove the mappings for now to reduce confusion when reading vimscript code with "normal!".